+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 - Added foreign key search back to DBC pages.
 - Added label search to DBC pages (clicking LabelID column values also searches).
 - Added reset button to all search modals on DBC pages.
-- DB2 search result modals now have tooltips/clickable columns where applicable.
-- DB2 search result modals now show the amount of results in tab headers.
+- Added tooltips/clickable columns to DB2 search results where applicable.
+- Added the amount of results in DB2 search result tab headers.
+- Added Creature/Quest names to tooltips (retrieved from local information or from archavon.kruithne.net)
 - Updated pagination on builds page to use input pagination.
 - Updated TACT key metadata.
 - Fixed diff cache not being reset after scanning unknown files.